+++ test/surface/components/live_redirect_test.exs
@@ -27,7 +27,9 @@ defmodule Surface.Components.LiveRedirectTest do
         """
       end
 
-    assert html =~ actual_content("user", to: "/users/1")
+    assert html == """
+           <a href="/users/1" data-phx-link="redirect" data-phx-link-state="push">user</a>
+           """
   end
 
   test "creates a link without label" do
@@ -38,7 +40,9 @@ defmodule Surface.Components.LiveRedirectTest do
         """
       end
 
-    assert html =~ actual_content(to: "/users/1")
+    assert html == """
+           <a href="/users/1" data-phx-link="redirect" data-phx-link-state="push"></a>
+           """
   end
 
   test "creates a link with default slot" do
@@ -49,7 +53,9 @@ defmodule Surface.Components.LiveRedirectTest do
         """
       end
 
-    assert html =~ actual_content({:safe, "<span>user</span>"}, to: "/users/1")
+    assert html == """
+           <a href="/users/1" data-phx-link="redirect" data-phx-link-state="push"><span>user</span></a>
+           """
   end
 
   test "setting the class" do
@@ -60,7 +66,9 @@ defmodule Surface.Components.LiveRedirectTest do
         """
       end
 
-    assert html =~ actual_content("user", to: "/users/1", class: "link")
+    assert html == """
+           <a href="/users/1" class="link" data-phx-link="redirect" data-phx-link-state="push">user</a>
+           """
   end
 
   test "setting multiple classes" do
@@ -71,7 +79,9 @@ defmodule Surface.Components.LiveRedirectTest do
         """
       end
 
-    assert html =~ actual_content("user", to: "/users/1", class: "link primary")
+    assert html == """
+           <a href="/users/1" class="link primary" data-phx-link="redirect" data-phx-link-state="push">user</a>
+           """
   end
 
   test "passing other options" do
@@ -87,14 +97,9 @@ defmodule Surface.Components.LiveRedirectTest do
         """
       end
 
-    actual =
-      actual_content("user",
-        to: "/users/1",
-        class: "link",
-        method: :delete,
-        data: [confirm: "Really?"],
-        "csrf-token": "token"
-      )
+    actual = """
+    <a href="/users/1" class="link" method="delete" data-confirm="Really?" data-phx-link="redirect" data-phx-link-state="push" csrf-token="token">user</a>
+    """
 
     assert attr_map(html) == attr_map(actual)
   end
@@ -103,16 +108,5 @@ defmodule Surface.Components.LiveRedirectTest do
     [{_, attrs, _}] = Floki.parse_fragment!(html)
 
     Map.new(attrs)
-  end
-
-  defp actual_content(text, opts) do
-    text
-    |> Phoenix.LiveView.Helpers.live_redirect(opts)
-    |> Phoenix.HTML.html_escape()
-    |> Phoenix.HTML.safe_to_string()
-  end
-
-  defp actual_content(opts) do
-    actual_content("", opts)
   end
 end
